9 - refactor admin.go: split up to several files
