+++ controller/admin_submission.go
@@ -1,0 +1,95 @@
+package controller
+
+import (
+	"encoding/json"
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/model"
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/shared/view"
+	"log"
+	"net/http"
+)
+
+// AdminSubmissionGET handles GET-request to /admin/submission
+func AdminSubmissionGET(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+
+	var subRepo = model.SubmissionRepository{}
+
+	submissions, err := subRepo.GetAll()
+	if err != nil {
+		log.Println(err)
+		return
+	}
+
+	v := view.New(r)
+	v.Name = "admin/submission/index"
+
+	v.Vars["Submissions"] = submissions
+
+	v.Render(w)
+}
+
+// AdminSubmissionCreateGET handles GET-request to /admin/submission/create
+func AdminSubmissionCreateGET(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+	w.WriteHeader(http.StatusOK)
+
+	v := view.New(r)
+	v.Name = "admin/submission/create"
+
+	v.Render(w)
+}
+
+// AdminSubmissionCreatePOST handles POST-request to /admin/submission/create
+func AdminSubmissionCreatePOST(w http.ResponseWriter, r *http.Request) {
+	// Get data from the form
+	data := r.FormValue("data")
+	// Declare Form-struct
+	var form = model.Form{}
+	// Unmarshal the JSON-string sent from the form
+	err := json.Unmarshal([]byte(data), &form)
+	if err != nil {
+		log.Println(err)
+		return
+	}
+	// Declare empty slice for error messages
+	var errorMessages []string
+
+	// Check form name
+	if form.Name == "" {
+		errorMessages = append(errorMessages, "Form name cannot be blank.")
+	}
+
+	// Check number of fields
+	if len(form.Fields) == 0 {
+		errorMessages = append(errorMessages, "Form needs to have at least 1 field.")
+	}
+
+	// Check if any error messages has been appended
+	if len(errorMessages) != 0 {
+		// TODO (Svein): Keep data from the previous submit
+		w.Header().Set("Content-Type", "text/html; charset=utf-8")
+		w.WriteHeader(http.StatusOK)
+
+		v := view.New(r)
+		v.Name = "admin/submission/create"
+
+		v.Vars["Errors"] = errorMessages
+
+		v.Render(w)
+
+		return
+	}
+
+	// Declare an empty Repository for Submission
+	var repo = model.SubmissionRepository{}
+	// Insert data to database
+	err = repo.Insert(form)
+	if err != nil {
+		log.Println(err)
+		return
+	}
+
+	// Redirect to /admin/submission
+	http.Redirect(w, r, "/admin/submission", http.StatusFound)
+}
