+++ template/admin/submission/index.tmpl
@@ -10,8 +10,8 @@
 
 <div class="row">
     {{range .Submissions}}
-    <div class="col-4">
-        <div class="card">
+    <div class="col-3">
+        <div class="card shadow-sm mb-3">
             <div class="card-body">
                 <h5 class="card-title">{{.Form.Name}}</h5>
                 <h6 class="card-subtitle mb-2 text-muted">{{.Form.Created}}</h6>
